+++ src/lib/components/forms/dynamic-form/service/terra-dynamic-form.service.ts
@@ -1,9 +1,6 @@
 /* eslint-disable */
 // Adding disable only because of the experimental state of the service!
 import { Injectable } from '@angular/core';
-import { TerraBaseService } from '../../../../service/terra-base.service';
-import { TerraLoadingSpinnerService } from '../../../loading-spinner/service/terra-loading-spinner.service';
-import { Http } from '@angular/http';
 import {
     Observable,
     of
@@ -15,14 +12,8 @@ import {
 @Injectable({
     providedIn: 'root'
 })
-export class TerraDynamicFormService extends TerraBaseService
+export class TerraDynamicFormService
 {
-    constructor(private spinnerService:TerraLoadingSpinnerService,
-                private baseHttp:Http)
-    {
-        super(spinnerService, baseHttp, '', false);
-    }
-
     /**
      * @param data
      * @param url