+++ src/lib/components/forms/input/file-input/terra-file-input.component.spec.ts
@@ -1,4 +1,3 @@
-import { HttpModule } from '@angular/http';
 import { HttpClientModule } from '@angular/common/http';
 import {
     FormsModule,
@@ -77,7 +76,6 @@ describe('TerraFileInputComponent', () =>
                 ModalModule.forRoot(),
                 FormsModule,
                 ReactiveFormsModule,
-                HttpModule,
                 HttpClientModule,
                 LocalizationModule.forRoot(l10nConfig)
             ],