+++ src/lib/components/layouts/info-box/terra-info-box.component.ts
@@ -36,7 +36,7 @@ export class TerraInfoBoxComponent
     @Input()
     public inputNoWordBreak:boolean;
 
-    @ViewChild('buttonRef')
+    @ViewChild('buttonRef', { static: true })
     public _buttonRef:ElementRef;
 
     public get hasButtons():boolean