+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Terra-Components now complies with the angular package format and code autocompl
 * **terra-date-picker** deprecated. Use [angular material datepicker](https://material.angular.io/components/datepicker/overview) instead.
 * **terra-multi-split-view** removed deprecated terra-multi-split-view. See [SplitView migration guide](https://developers.plentymarkets.com/dev-doc/split-view-migration-guide) for upgrade.
 * **TerraBaseStorageService** removed extension to TerraBaseService. Instead make use of the new HttpClient in all derived classes (e.g. TerraFrontendStorageService). 
+* **http** removed deprecated TerraBaseService and @angular/http dependency. See [TerraBaseService migration guide](https://developers.plentymarkets.com/dev-doc/migration-tbs-to-http-client) for instructions to replace it.
+* **dependencies** Updated [angular](https://github.com/angular/angular) and [angular-l10n](https://github.com/robisim74/angular-l10n) to v8. Please visit https://update.angular.io to figure out how to update your app to angular v8.
 
 ### Fix
 * **terra-base-editor** removed unnecessary quill style import
@@ -27,6 +29,13 @@ Terra-Components now complies with the angular package format and code autocompl
 * **terra-form**
     - fixed validation issues caused by invisible form fields.
     - fixed inconsistency when treating the value of invisible form fields.
+* **terra-file-browser** resolved circular dependency between _terra-file-browser.component.ts_ and _file-list.component.ts_.
+
+<a name="4.9.0"></a>
+# 4.9.0 (29.01.2020)
+
+### Features
+* **terra-syntax-editor** is now deprecated. Use [ckEditor](https://github.com/ckeditor/ckeditor4-angular) instead.
 
 <a name="4.8.2"></a>
 # 4.8.2 (08.01.2020)